Fix compilation warnings in offsetoff macro

+++ list.h
@@ -37,7 +37,7 @@
 /**
  * Get offset of a member
  */
-#define offsetof(TYPE, MEMBER) ((size_t) &((TYPE *)0)->MEMBER)
+#define offsetof(TYPE, MEMBER) ((size_t) ((unsigned long)&((TYPE *)0)->MEMBER))
 #endif
 
 /**
